Add files via upload
One of the central challenge for extracting governing principles of dynamical system via Dynamic Mode Decomposition (DMD) is about the limit data availability or
formally called as limited data acquisition in the present paper. In the interest of discovering the governing principles for a dynamical system with limited data acquisition, we provide a variant of Kernelized Extended DMD (KeDMD) based on the Koopman operator which employ the notion of Gaussian random matrix to recover the dominant Koopman modes for the standard fluid flow across cylinder experiment. It turns out that the traditional kernel function, Gaussian Radial Basis Function Kernel, unfortunately, is not able to generate the desired Koopman modes in the scenario of executing KeDMD with limited data acquisition. However, the Laplacian Kernel Function successfully generates the desired Koopman modes when limited data is provided in terms of data-set snapshot for the aforementioned experiment and this manuscripts serves the purpose of reporting these exciting experimental insights. This paper also explores the functionality of the Koopman operator when it interacts with the reproducing kernel Hilbert space (RKHS) that arises from the normalized probability Lebesgue measure dµσ,1,Cn (z) := (2πσ2)^−n exp (−∥z∥2/σ) dV (z) when it is embedded in L2−sense for the holomorphic functions over C^n, in the aim of determining the Koopman modes for fluid flow across cylinder experiment. We explore the operator-theoretic characterizations of the Koopman operator on the RKHS generated by the normalized Laplacian measure dµσ,1,Cn (z) in the L2−sense. In doing so, we provide the compactification & closable characterization of Koopman operator over the RKHS generated by the normalized Laplacian measure in the L2−sense.